refactor: extract column detection
